# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation jardin de cocagne",
        "version": "0.1"
    },
    "paths": {
...
        "/api/depot": {
            "post": {
                "tags": [
                    "Depots"
                ],
                "summary": "Add a new depository",
                "description": "Add a new depository for a structure",
                "operationId": "addNewDepository",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_structure",
                                    "name",
                                    "city",
                                    "address",
                                    "zip_code",
                                    "phone",
                                    "mail",
                                    "web_site",
                                    "delivery_day",
                                    "delivery_hour",
                                    "basket_hour",
                                    "presentation_text",
                                    "image_path",
                                    "comment"
                                ],
                                "properties": {
                                    "id_structure": {
                                        "type": "numeric"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "mail": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "web_site": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "delivery_day": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "delivery_hour": {
                                        "type": "date_format:H:i"
                                    },
                                    "basket_hour": {
                                        "type": "date_format:H:i"
                                    },
                                    "presentation_text": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "image_path": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "comment": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/depot/{id}": {
            "get": {
                "tags": [
                    "Depots"
                ],
                "summary": "Get a depository",
                "description": "Get a depository by ID",
                "operationId": "getDepositoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the depository to return",
                        "required": true,
                        "schema": {
                            "type": "numeric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                }
            },
            "put": {
                "tags": [
                    "Depots"
                ],
                "summary": "Update a depository",
                "description": "Update details of a depository by ID",
                "operationId": "updateDepository",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the depository to update",
                        "required": true,
                        "schema": {
                            "type": "numeric"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_structure",
                                    "name",
                                    "city",
                                    "address",
                                    "zip_code",
                                    "phone",
                                    "mail",
                                    "web_site",
                                    "delivery_day",
                                    "delivery_hour",
                                    "basket_hour",
                                    "presentation_text",
                                    "image_path",
                                    "comment"
                                ],
                                "properties": {
                                    "id_structure": {
                                        "type": "numeric"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "mail": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "web_site": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "delivery_day": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "delivery_hour": {
                                        "type": "date_format:H:i"
                                    },
                                    "basket_hour": {
                                        "type": "date_format:H:i"
                                    },
                                    "presentation_text": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "image_path": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "comment": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Depots"
                ],
                "summary": "Delete a depository",
                "description": "Delete a depository by ID",
                "operationId": "deleteDepository",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the depository to delete",
                        "required": true,
                        "schema": {
                            "type": "numeric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/depots": {
            "get": {
                "tags": [
                    "Depots"
                ],
                "summary": "Get all depositories",
                "description": "Get all depositories",
                "operationId": "getAllDepositories",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Operation failed"
                    }
                }
            }
        },
...
    },
...
}